requirements.txt
uvicorn
fastapi
python-multipart
tensorflow
keras
numpy
pandas
pillow